code/game/requirements.txt
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
Farama-Notifications==0.0.4
fonttools==4.54.1
gymnasium==0.29.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
packaging==24.1
pillow==10.4.0
pygame==2.6.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
typing_extensions==4.12.2
zipp==3.20.2